Update book model, breakup showBottomSheet
